+++ tests/test_whitebox.py
@@ -1,12 +1,13 @@
+import logging
+
 import numpy as np
 import pytest
-from time import time
 
-import logging
 logging.basicConfig(level=logging.DEBUG)
 
 try:
     import Pyro4
+
     skip_container_test = False
 except ImportError:
     skip_container_test = True
@@ -23,11 +24,12 @@ def test_whitebox_without_container_xgb():
 
     n_estimator = 32
     subsample = 1
-    result_dict = b.objective_function(configuration, n_estimators=n_estimator, subsample=subsample, rng=0)
+    result_dict = b.objective_function(configuration, fidelity=dict(n_estimators=n_estimator, subsample=subsample),
+                                       rng=0)
     valid_loss = result_dict['function_value']
-    train_loss = result_dict['train_loss']
+    train_loss = result_dict['info']['train_loss']
 
-    result_dict = b.objective_function_test(configuration, n_estimators=n_estimator, rng=0)
+    result_dict = b.objective_function_test(configuration, fidelity=dict(n_estimators=n_estimator), rng=0)
     test_loss = result_dict['function_value']
 
     assert np.isclose(train_loss, 0.1071, atol=0.001)
@@ -49,10 +51,10 @@ def test_whitebox_with_container():
 
     n_estimator = 32
     subsample = 1
-    result_dict = b.objective_function(configuration, n_estimators=n_estimator, subsample=subsample)
+    result_dict = b.objective_function(configuration, fidelity=dict(n_estimators=n_estimator, subsample=subsample))
     valid_loss = result_dict['function_value']
-    train_loss = result_dict['train_loss']
-    result_dict = b.objective_function_test(configuration, n_estimators=n_estimator)
+    train_loss = result_dict['info']['train_loss']
+    result_dict = b.objective_function_test(configuration, fidelity=dict(n_estimators=n_estimator))
     test_loss = result_dict['function_value']
 
     print(train_loss, valid_loss, test_loss)